Create README.md
In this code, we use the prompt() function, which opens a dialog box where the user can enter their name. We then store the entered name in the name variable. Next, we create a message in which we use the value of the name variable and save it to the message variable. Finally, we use the alert() function to display a welcome message with the username.

You can add this code to an HTML document inside a <script> tag, or you can execute it in the browser console. Upon launch, you will be prompted to enter your name and then a welcome message will appear with your name.